bmp image loading

+++ src/Image.zig
@@ -3,8 +3,8 @@ const std = @import("std");
 
 allocator: std.mem.Allocator,
 pixels: []u32,
-width: u64,
-height: u64,
+width: u32,
+height: u32,
 bpp: u16,
 
 pub fn deinit(self: *const Self) void {
